+++ crates/module-system/sov-modules-api/src/transaction/mod.rs
@@ -3,11 +3,13 @@ mod rewards;
 use std::fmt::Debug;
 use std::io;
 
+use crate::Multisig;
 use borsh::{BorshDeserialize, BorshSerialize};
 pub use data::{AuthenticatedTransactionData, Credentials, PriorityFeeBips, TxDetails};
 pub(crate) use rewards::transaction_consumption_helper;
 pub use rewards::{ProverReward, RemainingFunds, SequencerReward, TransactionConsumption};
 use serde::{Deserialize, Serialize};
+use sov_rollup_interface::common::SafeVec;
 #[cfg(feature = "native")]
 pub use sov_rollup_interface::crypto::PrivateKey;
 use sov_rollup_interface::crypto::SigVerificationError;
@@ -25,6 +27,9 @@ use crate::{
 
 #[cfg(test)]
 mod tests;
+
+/// The maximum number of signers allowed in a multisig.
+pub const MAX_SIGNERS: usize = 21;
 
 /// Structures that implement this trait represent a call message that can be included in a
 /// transaction.
@@ -78,12 +83,147 @@ pub struct Version0<Call, S: Spec> {
     serde::Serialize,
     serde::Deserialize,
     UniversalWallet,
+    PartialEq,
+    Eq,
+)]
+#[serde(bound = "S: Spec")]
+/// A signature and public key pair.
+pub struct PubKeyAndSignature<S: Spec> {
+    /// The signature.
+    pub signature: <S::CryptoSpec as CryptoSpec>::Signature,
+    /// The public key
+    pub pub_key: <S::CryptoSpec as CryptoSpec>::PublicKey,
+}
+
+impl<S: Spec> PubKeyAndSignature<S> {
+    /// Returns a reference to the public key.
+    pub fn key(&self) -> &<S::CryptoSpec as CryptoSpec>::PublicKey {
+        &self.pub_key
+    }
+}
+
+#[derive(
+    derive_more::Debug,
+    Clone,
+    borsh::BorshDeserialize,
+    serde::Serialize,
+    serde::Deserialize,
+    borsh::BorshSerialize,
+    UniversalWallet,
+)]
+#[serde(bound = "Call: serde::Serialize + serde::de::DeserializeOwned")]
+/// A V1 (multisig) transaction. The number of signers is capped at 10.
+///
+/// The credential ID for a multisig is hash(borsh(min_signers) || borsh(sort(pub_keys)))
+pub struct Version1<Call, S: Spec> {
+    /// The signatures of the transaction.
+    pub signatures: SafeVec<PubKeyAndSignature<S>, MAX_SIGNERS>,
+    /// The credential IDs that are part of the multisig but not used to sign the transaction.
+    // This is used to compute the credential ID statelessly
+    pub unused_pub_keys: SafeVec<<S::CryptoSpec as CryptoSpec>::PublicKey, MAX_SIGNERS>,
+    /// The minimum number of signers required to sign the transaction. In a 3/5 multisig, this would be 3.
+    /// Note that...
+    ///  - The transaction will be valid if at least `min_signers` signers have signed the transaction, but more signers are allowed. (This is useful for record keeping in case of, say, a unanimous decision)
+    ///  - If an invalid signature is detected, the transaction will be invalid. This allows us to use batch verification.
+    // This is used to compute the credential ID statelessly.
+    pub min_signers: u8,
+    /// The runtime call of the transaction.
+    #[sov_wallet(
+        bound = "Call: sov_rollup_interface::sov_universal_wallet::schema::UniversalWallet"
+    )]
+    pub runtime_call: Call,
+    /// Uniqueness identifier of this transaction. see [`UniquenessData`] for more details.
+    pub uniqueness: UniquenessData,
+    /// The transaction metadata. Contains gas parameters and the chain ID.
+    pub details: TxDetails<S>,
+}
+
+impl<Call: BorshSerialize, S: Spec> Version1<Call, S> {
+    /// Signs the transaction with the given key but does not add the signature to the list in the transaction.
+    #[cfg(feature = "native")]
+    pub fn sign_without_adding(
+        &self,
+        key: &<S::CryptoSpec as CryptoSpec>::PrivateKey,
+        chain_hash: &[u8; 32],
+    ) -> <S::CryptoSpec as CryptoSpec>::Signature {
+        key.sign(&self.serialize_for_signing(chain_hash))
+    }
+
+    /// Signs and adds the signature to the transaction.
+    #[cfg(feature = "native")]
+    pub fn sign(
+        &mut self,
+        key: &<S::CryptoSpec as CryptoSpec>::PrivateKey,
+        chain_hash: &[u8; 32],
+    ) -> anyhow::Result<()> {
+        let signature = self.sign_without_adding(key, chain_hash);
+        self.add_signature(signature, key.pub_key())
+    }
+
+    /// Serializes only the `UnsignedTransaction` part of the transaction and appens the chain_hash
+    pub fn serialize_for_signing(&self, chain_hash: &[u8; 32]) -> Vec<u8> {
+        let mut out = Vec::with_capacity(64); // Preallocate a little capacity to avoid excessive reallocations
+        BorshSerialize::serialize(&self.runtime_call, &mut out)
+            .expect("Serialization to vec is infallible");
+        BorshSerialize::serialize(&self.uniqueness, &mut out)
+            .expect("Serialization to vec is infallible");
+        BorshSerialize::serialize(&self.details, &mut out)
+            .expect("Serialization to vec is infallible");
+        out.extend_from_slice(chain_hash);
+        out
+    }
+
+    /// Adds a signature to the signing set of the multisig, removing the public key from the set of unused pub keys.
+    pub fn add_signature(
+        &mut self,
+        signature: <S::CryptoSpec as CryptoSpec>::Signature,
+        pub_key: <S::CryptoSpec as CryptoSpec>::PublicKey,
+    ) -> anyhow::Result<()> {
+        use crate::prelude::anyhow::Context;
+        if let Some(index) = self.unused_pub_keys.iter().position(|k| k == &pub_key) {
+            self.signatures
+                .try_push(PubKeyAndSignature { signature, pub_key })
+                .context("Too many signatures in multisig")?;
+            self.unused_pub_keys.remove(index);
+            Ok(())
+        } else {
+            anyhow::bail!("Public key is not a member of the multisig or has already signed");
+        }
+    }
+
+    /// Checks if enough signers have signed the transaction for it to be valid.
+    pub fn is_fully_signed(&self) -> bool {
+        self.signatures.len() >= self.min_signers as usize
+    }
+}
+
+impl<R: TransactionCallable, S: Spec> From<Version0<R::Call, S>> for Transaction<R, S> {
+    fn from(value: Version0<R::Call, S>) -> Self {
+        Transaction::V0(value)
+    }
+}
+
+impl<R: TransactionCallable, S: Spec> From<Version1<R::Call, S>> for Transaction<R, S> {
+    fn from(value: Version1<R::Call, S>) -> Self {
+        Transaction::V1(value)
+    }
+}
+#[derive(
+    derive_more::Debug, // derive_more uses the correct bound of TransactionCallable::RuntimeCall
+    Clone,
+    borsh::BorshSerialize,
+    borsh::BorshDeserialize,
+    serde::Serialize,
+    serde::Deserialize,
+    UniversalWallet,
 )]
 #[serde(bound = "R::Call: serde::Serialize + serde::de::DeserializeOwned")]
 /// A Transaction object that is compatible with the module-system/sov-default-stf.
 pub enum Transaction<R: TransactionCallable, S: Spec> {
     /// V0 Transaction type.
     V0(Version0<R::Call, S>),
+    /// A V1 (multisig) transaction.
+    V1(Version1<R::Call, S>),
 }
 
 #[cfg(feature = "native")]
@@ -170,6 +310,13 @@ impl<R: TransactionCallable, S: Spec> PartialEq for Transaction<R, S> {
                     && self_inner.uniqueness == other_inner.uniqueness
                     && self_inner.details == other_inner.details
             }
+            (Transaction::V1(self_inner), Transaction::V1(other_inner)) => {
+                self_inner.signatures == other_inner.signatures
+                    && self_inner.runtime_call == other_inner.runtime_call
+                    && self_inner.uniqueness == other_inner.uniqueness
+                    && self_inner.details == other_inner.details
+            }
+            _ => false,
         }
     }
 }
@@ -208,6 +355,7 @@ impl<R: TransactionCallable, S: Spec> Transaction<R, S> {
     pub fn runtime_call(&self) -> &R::Call {
         match &self {
             Transaction::V0(inner) => &inner.runtime_call,
+            Transaction::V1(inner) => &inner.runtime_call,
         }
     }
 
@@ -215,6 +363,7 @@ impl<R: TransactionCallable, S: Spec> Transaction<R, S> {
     pub fn chain_id(&self) -> u64 {
         match &self {
             Transaction::V0(inner) => inner.details.chain_id,
+            Transaction::V1(inner) => inner.details.chain_id,
         }
     }
 
@@ -259,6 +408,7 @@ impl<R: TransactionCallable, S: Spec> Transaction<R, S> {
     pub fn call(self) -> R::Call {
         match self {
             Transaction::V0(inner) => inner.runtime_call,
+            Transaction::V1(inner) => inner.runtime_call,
         }
     }
 
@@ -274,6 +424,24 @@ impl<R: TransactionCallable, S: Spec> Transaction<R, S> {
                     .verify(&inner.pub_key, msg, meter)
                     .map_err(TransactionVerificationError::from)?;
             }
+            Transaction::V1(inner) => {
+                for signature in inner.signatures.iter() {
+                    // Charge gas for all the signatures up front before verifying. This way, we can switch to batch verification and the gas price will be the same.
+                    MeteredSignature::new::<S>(signature.signature.clone())
+                        .charge_gas(meter, msg)
+                        .map_err(TransactionVerificationError::from)?;
+                }
+                let all_signers = inner
+                    .signatures
+                    .iter()
+                    .map(|s| s.pub_key.clone())
+                    .chain(inner.unused_pub_keys.iter().cloned())
+                    .collect::<Vec<_>>();
+                let multisig = Multisig::new(inner.min_signers, all_signers);
+                multisig
+                    .verify_signature(msg, &inner.signatures)
+                    .map_err(TransactionVerificationError::BadSignature)?;
+            }
         }
         Ok(())
     }
@@ -282,6 +450,11 @@ impl<R: TransactionCallable, S: Spec> Transaction<R, S> {
     pub fn to_unsigned_transaction(&self) -> UnsignedTransaction<R, S> {
         match &self {
             Transaction::V0(inner) => UnsignedTransaction::new_with_details(
+                inner.runtime_call.clone(),
+                inner.uniqueness,
+                inner.details.clone(),
+            ),
+            Transaction::V1(inner) => UnsignedTransaction::new_with_details(
                 inner.runtime_call.clone(),
                 inner.uniqueness,
                 inner.details.clone(),
@@ -363,6 +536,24 @@ impl<R: TransactionCallable, S: Spec> UnsignedTransaction<R, S> {
             self.uniqueness,
             self.details,
         )
+    }
+
+    /// Creates a new `V1` transaction from this unsigned transaction.
+    pub fn to_multisig_tx(
+        self,
+        multisig: Multisig<<S::CryptoSpec as CryptoSpec>::PublicKey>,
+    ) -> Version1<R::Call, S> {
+        Version1 {
+            signatures: SafeVec::new(),
+            unused_pub_keys: multisig
+                .signers
+                .try_into()
+                .expect("Too many signers in multisig"),
+            min_signers: multisig.required_signers,
+            runtime_call: self.runtime_call,
+            uniqueness: self.uniqueness,
+            details: self.details,
+        }
     }
 
     /// Returns a copy of the RuntimeCall from this unsigned transaction.